backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
    # via django
backports-zoneinfo==0.2.1
    # via django
black==22.10.0
    # via -r requirements.in
click==8.1.3
    # via black
django==4.1.3
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
django-environ==0.9.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
flake8==5.0.4
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
pathspec==0.10.1
    # via black
pillow==9.3.0
    # via -r requirements.in
platformdirs==2.5.3
    # via black
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2022.6
    # via djangorestframework
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via black
typing-extensions==4.4.0
    # via black
